Refactoring old tests

+++ tests/test_export.py
@@ -2,25 +2,38 @@ import os
 from os.path import isfile
 
 
-def test_export(sub):
-    # Check output_dialogues
+def test_export_output_dialogues(sub):
     assert sub.export(output_dialogues=True)
     assert not isfile("tests/sub.srt")
 
-    # Checking the default export
-    sub.export()
-    assert isfile("tests/sub.srt")
-    os.remove("tests/sub.srt")
 
-    # Checking output_dir
-    sub.export("tests/folder")
-    assert isfile("tests/folder/sub.srt")
-    os.remove("tests/folder/sub.srt")
+def test_export_default(sub):
+    try:
+        sub.export()
+        assert isfile("tests/sub.srt")
+    finally:
+        if isfile("tests/sub.srt"):
+            os.remove("tests/sub.srt")
 
-    # Checking compatibility
-    sub.export("tests/folder", "utf8", False)
-    assert isfile("tests/folder/sub.srt")
-    os.remove("tests/folder/sub.srt")
 
+def test_export_output_dir(sub):
+    try:
+        sub.export("tests/folder")
+        assert isfile("tests/folder/sub.srt")
+    finally:
+        if isfile("tests/folder/sub.srt"):
+            os.remove("tests/folder/sub.srt")
+
+
+def test_export_compatibility(sub):
+    try:
+        sub.export("tests/folder", "utf8", False)
+        assert isfile("tests/folder/sub.srt")
+    finally:
+        if isfile("tests/folder/sub.srt"):
+            os.remove("tests/folder/sub.srt")
+
+
+def test_export_output_dialogues_true(sub):
     assert sub.export(None, "utf8", True)
     assert not isfile("tests/sub.srt")
